+++ src/EnrolmentWorkflow/Action/SendNotification/SendNotificationFormGUI.php
@@ -6,6 +6,7 @@ use ilRadioGroupInputGUI;
 use ilRadioOption;
 use ilSelectInputGUI;
 use srag\CustomInputGUIs\SrUserEnrolment\MultiSelectSearchNewInputGUI\MultiSelectSearchNewInputGUI;
+use srag\CustomInputGUIs\SrUserEnrolment\MultiSelectSearchNewInputGUI\UsersAjaxAutoCompleteCtrl;
 use srag\Notifications4Plugin\SrUserEnrolment\Notification\NotificationInterface;
 use srag\Notifications4Plugin\SrUserEnrolment\Notification\NotificationsCtrl;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Action\AbstractActionFormGUI;
@@ -66,11 +67,10 @@ class SendNotificationFormGUI extends AbstractActionFormGUI
                             case SendNotification::TO_TYPE_SPECIFIC_USERS:
                                 $items = [
                                     "to_specific_users" => [
-                                        self::PROPERTY_CLASS    => MultiSelectSearchNewInputGUI::class,
-                                        self::PROPERTY_REQUIRED => true,
-                                        self::PROPERTY_OPTIONS  => self::srUserEnrolment()->ruleEnrolment()->searchUsers(),
-                                        "setAjaxLink"           => self::dic()->ctrl()->getLinkTarget($this->parent, ActionGUI::CMD_GET_USERS_AUTO_COMPLETE, "", true, false),
-                                        "setTitle"              => $this->txt("totype_specific_users")
+                                        self::PROPERTY_CLASS      => MultiSelectSearchNewInputGUI::class,
+                                        self::PROPERTY_REQUIRED   => true,
+                                        "setAjaxAutoCompleteCtrl" => new UsersAjaxAutoCompleteCtrl(),
+                                        "setTitle"                => $this->txt("totype_specific_users")
                                     ]
                                 ];
                                 break;